requirements.txt
requests
openai-agents
openai
python-dotenv